[WIP] Add the firsts controllers

+++ src/main/java/com/ravel/teste/srm/dto/ChangeResponseDTO.java
@@ -4,14 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ChangeResponseDTO {
-    private Double convertedAmount;
-    private Double taxRate;
+    private BigDecimal convertedAmount;
+    private BigDecimal taxRate;
     private String productName;
     private String originCoin;
     private String destinyCoin;
